Update angular dependencies, tell dependabot to ignore (#25)
Dependabot wasn't able to update any of the `@angular/*` dependencies because it was doing one at a time, causing cross-dependencies between them to get messed up. The solution is Angular's command `npm run ng update @angular/cli @angular/core`, which updates every Angular dependency at the same time. This should be run occasionally to make sure we're up to date, which also means we have to explicitly tell dependabot to ignore packages that start with `@angular/`.