Set up ServerApi utility in frontend (#38)
To help with the frontend side of the interactions between the frontend and backend, I added the ServerApi utility. It separates out some of the http getting logic, making it modular and unit-testable. Logic for making more http requests will be added in server-api.service.ts, and the unit tests for each one will be added in server-api.service.spec.ts.