Add projects to home page (#68)
The home page was previously just the boilerplate .NET + Angular stuff, so I removed that and added some cards to represent each project. I also renamed the component from `HomeComponent` to `HomePageComponent`.

As we begin to flesh out the other pages more, I also made the `TitleInplaceComponent`, which demonstrates how we want to implement inputs in our site. PrimeNG's `inplace` allows for the presentation of a value and its input to take up a single spot on the page. Right now I placed it on the project page, but I imagine it would be used on the task page too, at the very least. The format and setup of this inplace component should be useful for the other types of inplace components we'll make, like description, calendar, etc.