Project page buttons + sprint ordering (#75)
This was a quick set of changes that came from suggestions for extra functionality to the list of sprints on the project page. We can now collapse and expand every sprint with a single button push. By default, completed sprints will also be hidden from view, with a new checkbox button to reveal them and hide them again.

I also more rigidly defined the order in which sprints should appear on the page. Completed sprints should always be at the top, backlogs should always be at the bottom, and everything else should be ordered by start date, with oldest sprints at the top.

I also made sure the styling was more conducive to having multiple backlogs, previously it assumed there was just one.