Better home page routing (#79)
This is a quick one to make a home page that actually acts as a home page, being the main landing page that you go to before logging in. This means I had to move the functionality that used to be the `HomePageComponent` into the `ProjectsPageComponent`. `HomePageComponent` is now a basic page with the cuttlefish logo and a button to navigate to the login page. `ProjectsPageComponent` now shows the list of projects.

I also made it so that the nav bar doesn't show up until the user gets to the projects page, because users shouldn't be able to use it until they're logged in. I followed Hess's suggestions with the nav bar to make the list of projects show up in a dropdown, as well as make the logo at the top route to the projects dashboard.

The last tiny change was to update the favicon of the site to the cuttlefish logo.