Buttons & modals for creating entities (#86)
I created three new components, all very similar: `CreateProjectModalComponent`, `CreateSprintModalComponent`, and `CreateTaskModalComponent`. These will appear as a modal when buttons on the projects page, project page, and sprint dropdown are clicked, respectively. Once the modal inputs are confirmed, the new entity is created and shown on the page so that you can see the action taking place.

I didn't implement all the fields so that we can just get these started as a proof of concept so we're ready for the UI/UX team to begin using our site.

In the future, more fields need to be added to the modals, and we also need to verify the inputs by requiring fields, etc. I tracked those next steps here: https://trello.com/c/LYpaYOPu/62-finish-filling-out-create-entity-modals