Clean up the Log In and Sign Up pages (#104)
Pretty much all the functionality of logging in and signing up was already there thanks to Yev, so the majority of these changes are visual. Apart from that, I added more proper error catching when some typical login/signup errors are found, such as...
- invalid username or password on login
- username or email already taken on signup
To do that error catching and displaying the error in the frontend, I tweaked the backend a bit to respond with more specific error codes that I can check for in the frontend.

I also moved the signup password verification from the backend to the frontend to avoid extra requests to the backend. The frontend now verifies that the `password` and `confirmPassword` boxes...
- have the same value
- are at least 8 characters long
- have at least one capital letter
- have at least one special character
- have at least one number
Once all those are true, only then does it make a request to the backend, which checks for if the username and/or email are already taken. I also made sure to add unit testing for the function that checks the password.

Also, because I just discovered that you can use `dotnet watch run` to enable hot reloads, I added that to the readme.