Reorganize frontend file structure (#107)
Now, in `ClientApp/src/app`, instead of a massive list of folders of all types of entities, it's split up into `components/`, `guards/`, `interceptors/`, and `services/`. And inside `components/` I created a little bit more structure with different component types. I took this chance to clean up the imports too. Namely, switching all imports to be relative to `src/`. Imports that start with src aren't relative in the true sense, so they're more robust to changes. This means that from now on, every time we're importing stuff between .ts files, we should use import paths that start with `src/`. I think a valid exception to this is importing a class in it's corresponding .spec.ts file, I think it's fine to use a `./` import in that case because they're guaranteed to be right next to each other.

I also got rid of the old counter and fetch data pages that came with the ASP.NET & Angular boiler plate.