Update project creation (#136)
This was another set of small suggestions made by the ui/ux team. The first was to move the project creation button to the top of the page so that it doesn't get pushed to the bottom, which was a quick fix. The other was to add a "new project" button to the nav, in the list with all the other projects. And that was a pretty quick add as well.