Update styling for forgot/reset password pages (#135)
Not much additional content here, just matching the styling of the login and signup pages with the last two login-related pages. Mostly just switching stuff out for primeng components, with some smaller css tweaks.

The most significant thing I did was separate out our password verification functionality into its own service. This avoids too much duplicate code, since it only needed to be done on the signup page previously, but now would need to be done on the reset password page too. The service makes sure the `password` and `confirmpassword` are the same, as well as verifying the strength of the password.

The last small reorganization I did was the checking of error codes for the reset password functionality. I made it so the service is the one checking the error codes and deducing a reasonable error message from that, instead of the component being in charge of it.