Sprint completion and deletion (#157)
Here's the functionality for completing and deleting sprints. The biggest part of this is that when completing or deleting a sprint, it checks what the user wants to do with them and takes that action. The three options for sprint completion, if there are tasks in the sprint not marked as done yet, are to 1. move the tasks to an available sprint, 2. move the tasks to an available sprint, and 3. mark the tasks as complete.

![image](https://user-images.githubusercontent.com/60301051/234775398-581b0964-7cb2-4625-954f-27856aa2571c.png)

The next step after hitting continue is to generate and show the sprint report, which has a couple pie charts for task/point completion and a burndown of when the tasks were completed over the course of the sprint.

![image](https://user-images.githubusercontent.com/60301051/234776489-340f4c91-48c8-4c58-be79-b7fb4089f37a.png)

I tried primeng's charts at first, but there was some weirdness going on with chart.js, which primeng uses. I ended up using apexcharts instead, which is another modern option that had angular support.

Sprint deletion is mostly the same, it provides options for how to deal with the incomplete tasks within the sprint before deleting. They can move them to another task or backlog, or opt to delete them with the rest of the complete tasks in the sprint. I'm relying on nathan's cascading to delete a task when its sprint is deleted.

I also had a couple other small changes, including to the progress picker on the task page so that it makes changes in the backend properly. I also noticed that the task service wasn't properly converting the start and end dates from strings to date objects, so I updated it to convert from an ISO string to a date object.