Drag and drop tasks (#164)
Here's drag and drop! I used Angular CDK's `DragDropModule` to handle drag and drop to order tasks within sprints and move them between sprints. The core of this was adding a new `order` property to `TaskData` and keeping track of it in the database properly so that changes would persist. I added 3 new endpoints to the backend to manage the order updating in sprints, and created `SprintOrderingService` to hit those endpoints. I added the use of that service into the task service so that deleting or creating a new task also kept the ordering right.

The hardest part of this was implementing automatic scrolling when you're dragging a task and your cursor is close to the top or bottom of the page. It took some tweaking but I got it to somewhat work, the scrolling isn't perfectly smooth but it'll do the job.