Message service cleanup (#168)
I mostly just cleaned up most of the message service messages. If a user makes a successful change, we only want to display a `<p-toast>` message if you can't visually see the change happen in the UI, so I got rid of most of the success messages, keeping a couple of them around if a UI change doesn't happen in that case.

I also removed extra console.logs and added message service messages where we were previously logging the error. That required adding `MessageService` and `<p-toast>` to a couple components.

The last thing I did was tweak the error checking on the task page so that it redirects to the 404 page if the task wasn't found.