Better looking task progress tags (#171)
I created a very simple new component `TaskProgressTagComponent` to handle the progress, so every time you need to display the progress, we'll use `<task-progress-tag>`. Its HTML is just a `<p-tag>`, but it's set up to manage the colors and icons used by the progress tags:

I'm open to switching the colors and icons if you guys think something else would be a better fit. It'll be interesting to see if the ui/ux team agrees with them too.

The last thing I did was update a couple unit tests to fix warnings that came up from my last pr.