Task date edge case (#191)
There was an edge case we discovered where moving tasks between sprints didn't update the task's dates, making them invalid. I created a new function to perform that task date moving and wrote a couple unit tests for it. I also made sure the changelog button in the footer linked to the changelog and reverted the updates I made to `isTaskData()`.