Can now disable drag and drop (#192)
I added a new input to the `task-overview` to disable drag and drop on it, which removes the draggable handle on the left side of it. I ended up having to copy and paste some HTML because it's not possible to conditionally apply a directive, which the angular cdk uses to implement drag and drop. So unfortunately, the basic and kinda clunky solution is to use `*ngIf`.

I also did a little bit of cleanup in the heat of the moment, so now the `task-overview`'s description only has ellipsis if the description had to be cut off for being too long. Previously, it just had ellipsis all the time and looked misleading.